+++ tests/Tgstation.Server.Tests/Live/RateLimitRetryingApiClient.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+
+using Tgstation.Server.Api;
+using Tgstation.Server.Client;
+using Tgstation.Server.Common;
+
+namespace Tgstation.Server.Tests.Live
+{
+	sealed class RateLimitRetryingApiClient : ApiClient
+	{
+		public RateLimitRetryingApiClient(IHttpClient httpClient, Uri url, ApiHeaders apiHeaders, ApiHeaders tokenRefreshHeaders, bool authless)
+			: base(httpClient, url, apiHeaders, tokenRefreshHeaders, authless)
+		{
+		}
+
+		protected override async Task<TResult> RunRequest<TResult>(string route, HttpContent content, HttpMethod method, long? instanceId, bool tokenRefresh, CancellationToken cancellationToken)
+		{
+			var hasGitHubToken = !String.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("TGS_TEST_GITHUB_TOKEN"));
+			while (true)
+				try
+				{
+					return await base.RunRequest<TResult>(route, content, method, instanceId, tokenRefresh, cancellationToken);
+				}
+				catch (RateLimitException ex) when (hasGitHubToken && ex.RetryAfter.HasValue)
+				{
+					var now = DateTimeOffset.UtcNow;
+
+					Console.WriteLine($"TEST ERROR RATE LIMITED: {ex}");
+
+					var sleepTime = ex.RetryAfter.Value - now;
+					Console.WriteLine($"Sleeping for {sleepTime.TotalMinutes} minutes and retrying...");
+					await Task.Delay(sleepTime, cancellationToken);
+				}
+		}
+	}
+}
