+++ tests/Tgstation.Server.Tests/Live/TestLiveServer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -12,9 +11,6 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -29,8 +25,6 @@ using Tgstation.Server.Client;
 using Tgstation.Server.Client.Components;
 using Tgstation.Server.Host.Components;
 using Tgstation.Server.Host.Configuration;
-using Tgstation.Server.Host.Database;
-using Tgstation.Server.Host.Database.Migrations;
 using Tgstation.Server.Host.Extensions;
 using Tgstation.Server.Host.System;
 using Tgstation.Server.Tests.Live.Instance;
@@ -75,6 +69,12 @@ namespace Tgstation.Server.Tests.Live
 			{
 				l.Stop();
 			}
+		}
+
+		[TestInitialize]
+		public void Initialize()
+		{
+			ServerClientFactory.ApiClientFactory = new RateLimitRetryingApiClientFactory();
 		}
 
 		[TestMethod]
@@ -128,13 +128,6 @@ namespace Tgstation.Server.Tests.Live
 				var updatedAssemblyVersion = FileVersionInfo.GetVersionInfo(updatedAssemblyPath);
 				Assert.AreEqual(testUpdateVersion, Version.Parse(updatedAssemblyVersion.FileVersion).Semver());
 			}
-			catch (RateLimitException ex)
-			{
-				if (string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("TGS_TEST_GITHUB_TOKEN")))
-					throw;
-
-				Assert.Inconclusive("GitHub rate limit hit: {0}", ex);
-			}
 			finally
 			{
 				serverCts.Cancel();
@@ -170,7 +163,7 @@ namespace Tgstation.Server.Tests.Live
 			}, false, 5011))
 			{
 				using var serverCts = new CancellationTokenSource();
-				serverCts.CancelAfter(TimeSpan.FromMinutes(3));
+				serverCts.CancelAfter(TimeSpan.FromHours(3));
 				var cancellationToken = serverCts.Token;
 				var serverTask = controller.Run(cancellationToken);
 
@@ -216,13 +209,6 @@ namespace Tgstation.Server.Tests.Live
 					}
 
 					CheckServerUpdated(controller);
-				}
-				catch (RateLimitException ex)
-				{
-					if (string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("TGS_TEST_GITHUB_TOKEN")))
-						throw;
-
-					Assert.Inconclusive("GitHub rate limit hit: {0}", ex);
 				}
 				finally
 				{
@@ -452,13 +438,6 @@ namespace Tgstation.Server.Tests.Live
 					CheckServerUpdated(node1);
 					CheckServerUpdated(node2);
 				}
-				catch (RateLimitException ex)
-				{
-					if (string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("TGS_TEST_GITHUB_TOKEN")))
-						throw;
-
-					Assert.Inconclusive("GitHub rate limit hit: {0}", ex);
-				}
 				finally
 				{
 					serverCts.Cancel();
@@ -636,13 +615,6 @@ namespace Tgstation.Server.Tests.Live
 					Assert.AreEqual(2, node2Info.SwarmServers.Count);
 					Assert.IsNotNull(node2Info.SwarmServers.SingleOrDefault(x => x.Identifier == "controller"));
 				}
-				catch (RateLimitException ex)
-				{
-					if (string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("TGS_TEST_GITHUB_TOKEN")))
-						throw;
-
-					Assert.Inconclusive("GitHub rate limit hit: {0}", ex);
-				}
 				finally
 				{
 					serverCts.Cancel();
@@ -651,157 +623,6 @@ namespace Tgstation.Server.Tests.Live
 			}
 
 			new LiveTestingServer(null, false).Dispose();
-		}
-
-		[TestMethod]
-		public async Task TestDownMigrations()
-		{
-			var connectionString = Environment.GetEnvironmentVariable("TGS_TEST_CONNECTION_STRING");
-
-			if (string.IsNullOrEmpty(connectionString))
-				Assert.Inconclusive("No connection string configured in env var TGS_TEST_CONNECTION_STRING!");
-
-			var databaseTypeString = Environment.GetEnvironmentVariable("TGS_TEST_DATABASE_TYPE");
-			if (!Enum.TryParse<DatabaseType>(databaseTypeString, out var databaseType))
-				Assert.Inconclusive("No/invalid database type configured in env var TGS_TEST_DATABASE_TYPE!");
-
-			string migrationName = null;
-			DatabaseContext CreateContext()
-			{
-				string serverVersion = Environment.GetEnvironmentVariable($"{DatabaseConfiguration.Section}__{nameof(DatabaseConfiguration.ServerVersion)}");
-				if (string.IsNullOrWhiteSpace(serverVersion))
-					serverVersion = null;
-				switch (databaseType)
-				{
-					case DatabaseType.MySql:
-					case DatabaseType.MariaDB:
-						migrationName = nameof(MYInitialCreate);
-						return new MySqlDatabaseContext(
-							Host.Database.Design.DesignTimeDbContextFactoryHelpers.CreateDatabaseContextOptions<MySqlDatabaseContext>(
-								databaseType,
-								connectionString,
-								serverVersion));
-					case DatabaseType.PostgresSql:
-						migrationName = nameof(PGCreate);
-						return new PostgresSqlDatabaseContext(
-							Host.Database.Design.DesignTimeDbContextFactoryHelpers.CreateDatabaseContextOptions<PostgresSqlDatabaseContext>(
-								databaseType,
-								connectionString,
-								serverVersion));
-					case DatabaseType.SqlServer:
-						migrationName = nameof(MSInitialCreate);
-						return new SqlServerDatabaseContext(
-							Host.Database.Design.DesignTimeDbContextFactoryHelpers.CreateDatabaseContextOptions<SqlServerDatabaseContext>(
-								databaseType,
-								connectionString,
-								serverVersion));
-					case DatabaseType.Sqlite:
-						migrationName = nameof(SLRebuild);
-						return new SqliteDatabaseContext(
-							Host.Database.Design.DesignTimeDbContextFactoryHelpers.CreateDatabaseContextOptions<SqliteDatabaseContext>(
-								databaseType,
-								connectionString,
-								serverVersion));
-				}
-
-				return null;
-			}
-
-			using var context = CreateContext();
-			await context.Database.EnsureDeletedAsync();
-			await context.Database.MigrateAsync(default);
-
-			// add usergroups and dummy instances for testing purposes
-			var group = new Host.Models.UserGroup
-			{
-				PermissionSet = new Host.Models.PermissionSet
-				{
-					AdministrationRights = AdministrationRights.ChangeVersion,
-					InstanceManagerRights = InstanceManagerRights.GrantPermissions
-				},
-				Name = "TestGroup",
-			};
-
-			const string TestUserName = "TestUser42";
-			var user = new Host.Models.User
-			{
-				Name = TestUserName,
-				CreatedAt = DateTimeOffset.UtcNow,
-				OAuthConnections = new List<Host.Models.OAuthConnection>(),
-				CanonicalName = Host.Models.User.CanonicalizeName(TestUserName),
-				Enabled = false,
-				Group = group,
-				PasswordHash = "_",
-			};
-
-			var instance = new Host.Models.Instance
-			{
-				AutoUpdateInterval = 0,
-				ChatBotLimit = 1,
-				ChatSettings = new List<Host.Models.ChatBot>(),
-				ConfigurationType = ConfigurationType.HostWrite,
-				DreamDaemonSettings = new Host.Models.DreamDaemonSettings
-				{
-					AllowWebClient = false,
-					AutoStart = false,
-					HeartbeatSeconds = 0,
-					DumpOnHeartbeatRestart = false,
-					Port = 1447,
-					SecurityLevel = DreamDaemonSecurity.Safe,
-					Visibility = DreamDaemonVisibility.Public,
-					StartupTimeout = 1000,
-					TopicRequestTimeout = 1000,
-					AdditionalParameters = string.Empty,
-					StartProfiler = false,
-					LogOutput = true,
-				},
-				DreamMakerSettings = new Host.Models.DreamMakerSettings
-				{
-					ApiValidationPort = 1557,
-					ApiValidationSecurityLevel = DreamDaemonSecurity.Trusted,
-					RequireDMApiValidation = false,
-					Timeout = TimeSpan.FromSeconds(13),
-				},
-				InstancePermissionSets = new List<Host.Models.InstancePermissionSet>
-				{
-					new Host.Models.InstancePermissionSet
-					{
-						ByondRights = ByondRights.InstallCustomVersion,
-						ChatBotRights = ChatBotRights.None,
-						ConfigurationRights = ConfigurationRights.Read,
-						DreamDaemonRights = DreamDaemonRights.ReadRevision,
-						DreamMakerRights = DreamMakerRights.SetApiValidationPort,
-						InstancePermissionSetRights = InstancePermissionSetRights.Write,
-						PermissionSet = group.PermissionSet,
-						RepositoryRights = RepositoryRights.SetReference
-					}
-				},
-				Name = "sfdsadfsa",
-				Online = false,
-				Path = "/a/b/c/d",
-				RepositorySettings = new Host.Models.RepositorySettings
-				{
-					AutoUpdatesKeepTestMerges = false,
-					AutoUpdatesSynchronize = false,
-					CommitterEmail = "email@eample.com",
-					CommitterName = "blubluh",
-					CreateGitHubDeployments = false,
-					PostTestMergeComment = false,
-					PushTestMergeCommits = false,
-					ShowTestMergeCommitters = false,
-					UpdateSubmodules = false,
-				},
-			};
-
-			context.Users.Add(user);
-			context.Groups.Add(group);
-			context.Instances.Add(instance);
-			await context.Save(default);
-
-			var dbServiceProvider = ((IInfrastructure<IServiceProvider>)context.Database).Instance;
-			var migrator = dbServiceProvider.GetRequiredService<IMigrator>();
-			await migrator.MigrateAsync(migrationName, default);
-			await context.Database.EnsureDeletedAsync();
 		}
 
 		[TestMethod]
@@ -817,7 +638,7 @@ namespace Tgstation.Server.Tests.Live
 
 			using var server = new LiveTestingServer(null, true);
 
-			const int MaximumTestMinutes = 20;
+			const int MaximumTestMinutes = 180;
 			using var hardTimeoutCancellationTokenSource = new CancellationTokenSource(TimeSpan.FromMinutes(MaximumTestMinutes));
 			var hardCancellationToken = hardTimeoutCancellationTokenSource.Token;
 			using var serverCts = CancellationTokenSource.CreateLinkedTokenSource(hardCancellationToken);
@@ -1066,12 +887,12 @@ namespace Tgstation.Server.Tests.Live
 			}
 			catch (ApiException ex)
 			{
-				System.Console.WriteLine($"[{DateTimeOffset.UtcNow}] TEST ERROR: {ex.ErrorCode}: {ex.Message}\n{ex.AdditionalServerData}");
+				Console.WriteLine($"[{DateTimeOffset.UtcNow}] TEST ERROR: {ex.ErrorCode}: {ex.Message}\n{ex.AdditionalServerData}");
 				throw;
 			}
 			catch (Exception ex)
 			{
-				System.Console.WriteLine($"[{DateTimeOffset.UtcNow}] TEST ERROR: {ex}");
+				Console.WriteLine($"[{DateTimeOffset.UtcNow}] TEST ERROR: {ex}");
 				throw;
 			}
 			finally
